delete old thread reaction button, add styles to reaction button
