--- client/styles/components/thread_reaction_button.scss
@@ -1,78 +1,0 @@
-@import 'client/styles/shared';
-
-.ThreadReactionButton {
-  user-select: none;
-  display: flex;
-  .cui-icon,
-  .reply-count,
-  .reactions-count {
-    color: $black;
-    font-weight: 500;
-  }
-  .cui-icon {
-    margin-right: 8px;
-  }
-  .reply-count,
-  .reactions-count {
-    position: relative;
-    top: 0px;
-  }
-  // border
-  &.disabled {
-    background: initial !important;
-    color: $neutral-500 !important;
-    border-color: $neutral-50 !important;
-  }
-  &.large {
-    display: block;
-    min-height: 43px !important;
-    min-width: 45px !important;
-    max-height: 43px !important;
-    max-width: 45px !important;
-    text-align: center;
-    color: $neutral-500;
-    border: 1px solid $neutral-100;
-    display: block;
-    border-radius: 12px;
-    padding: 0;
-    &.active {
-      color: $primary-500;
-      border-color: $primary-500;
-      span {
-        color: inherit;
-      }
-    }
-    .cui-icon {
-      margin-right: none;
-    }
-    .reactions-icon {
-      position: relative;
-      top: 1px;
-      font-size: 85%;
-      padding: 0;
-      margin: 0;
-      margin-bottom: 4px;
-    }
-    &.like .reactions-icon {
-      -webkit-transform: rotate(180deg);
-      -moz-transform: rotate(180deg);
-      -o-transform: rotate(180deg);
-      -ms-transform: rotate(180deg);
-      transform: rotate(180deg);
-    }
-    .reactions-count {
-      top: 0;
-      padding: 0;
-      min-width: 26px;
-      text-align: center;
-      color: $neutral-500;
-      font-weight: 500;
-    }
-  }
-}
-
-.cui-popover.cui-tooltip.ReactionButtonTooltip {
-  .cui-popover-content .reaction-button-tooltip {
-    padding: 7px 8px 2px;
-  }
-}
